131 nas in the std values when parameters are present in the model (#132)
* fix: take into account sign when computing log det

arma::log_det returns a finite number even when the determinant is negative, because it computes log|det| plus a sign. The sign was so far discarded. This is hopefully the root cause of the error we were getting, i.e., that sometimes we had parameters in the model but standard deviations were turning up as NAs

* feat: use devtools::load_all() when regenarating data

* fix: regenerated data after the fix. The results for PIP, PM and PSDR have been compared with the original Moral-Benito results and they match almost perfectly (at most 0.001 difference for PM and PSDR and only in rare cases; PIP up to 1%, also in rare cases)

* test: add tests that check that stds are non-zero if a parameter is present in the model. Also check that likelihoods and BICs are always non zero

* fix: use appropriate object to check model space stats

* test: stds have to be greater than zero, not only different than zero