test: remove tests on the model space level
The tests fail even with tolerance 0.05. Therefore they make no sense. I'm not sure if we should expect optimisation in such a high dimensional space to give always the same results. For now we have to resort to low-level tests, and the test that checks that we can replicate Moral-Benito results.